// src/lib/signalement/models/ExistingVoie.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ExistingVoie = {
    type: ExistingVoie.type;
    nom: string;
};
export namespace ExistingVoie {
    export enum type {
        NUMERO = 'NUMERO',
        VOIE = 'VOIE',
        TOPONYME = 'TOPONYME',
    }
}